@@ -1,12 +1,12 @@
 class UserModel {
   String uid;
-  String email;
-  String displayName;
-  String phoneNumber;
-  String photoUrl;
+  String? email;
+  String? displayName;
+  String? phoneNumber;
+  String? photoUrl;
 
   UserModel(
-      {this.uid,
+      {required this.uid,
       this.email,
       this.displayName,
       this.phoneNumber,
