@@ -3,7 +3,7 @@ import 'package:noteapp/caches/sharedpref/shared_preference_helper.dart';
 
 class LanguageProvider extends ChangeNotifier {
   // shared pref object
-  SharedPreferenceHelper _sharedPrefsHelper;
+  late SharedPreferenceHelper _sharedPrefsHelper;
 
   Locale _appLocale = Locale('en');
 
@@ -12,10 +12,8 @@ class LanguageProvider extends ChangeNotifier {
   }
 
   Locale get appLocale {
-    _sharedPrefsHelper.appLocale.then((localeValue) {
-      if (localeValue != null) {
-        _appLocale = Locale(localeValue);
-      }
+    _sharedPrefsHelper.appLocale?.then((localeValue) {
+      _appLocale = Locale(localeValue);
     });
 
     return _appLocale;
