@@ -1,16 +1,16 @@
 import 'dart:async';
 
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:meta/meta.dart';
 import 'package:noteapp/models/todo_model.dart';
 import 'package:noteapp/services/firestore_path.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
+
 import 'package:noteapp/services/firestore_service.dart';
 
 String documentIdFromCurrentDate() => DateTime.now().toIso8601String();
 
 /*
 This is the main class access/call for any UI widgets that require to perform
-any CRUD activities operation in Firestore database.
+any CRUD activities operation in FirebaseFirestore database.
 This class work hand-in-hand with FirestoreService and FirestorePath.
 
 Notes:
@@ -21,13 +21,13 @@ changed to true.
 
  */
 class FirestoreDatabase {
-  FirestoreDatabase({@required this.uid}) : assert(uid != null);
+  FirestoreDatabase({required this.uid}) : assert(uid != null);
   final String uid;
 
   final _firestoreService = FirestoreService.instance;
 
   //Method to create/update todoModel
-  Future<void> setTodo(TodoModel todo) async => await _firestoreService.setData(
+  Future<void> setTodo(TodoModel todo) async => await _firestoreService.set(
         path: FirestorePath.todo(uid, todo.id),
         data: todo.toMap(),
       );
@@ -38,7 +38,7 @@ class FirestoreDatabase {
   }
 
   //Method to retrieve todoModel object based on the given todoId
-  Stream<TodoModel> todoStream({@required String todoId}) =>
+  Stream<TodoModel> todoStream({required String todoId}) =>
       _firestoreService.documentStream(
         path: FirestorePath.todo(uid, todoId),
         builder: (data, documentId) => TodoModel.fromMap(data, documentId),
@@ -52,27 +52,27 @@ class FirestoreDatabase {
 
   //Method to mark all todoModel to be complete
   Future<void> setAllTodoComplete() async {
-    final batchUpdate = Firestore.instance.batch();
+    final batchUpdate = FirebaseFirestore.instance.batch();
 
-    final querySnapshot = await Firestore.instance
+    final querySnapshot = await FirebaseFirestore.instance
         .collection(FirestorePath.todos(uid))
-        .getDocuments();
+        .get();
 
-    for (DocumentSnapshot ds in querySnapshot.documents) {
-      batchUpdate.updateData(ds.reference, {'complete': true});
+    for (DocumentSnapshot ds in querySnapshot.docs) {
+      batchUpdate.update(ds.reference, {'complete': true});
     }
     await batchUpdate.commit();
   }
 
   Future<void> deleteAllTodoWithComplete() async {
-    final batchDelete = Firestore.instance.batch();
+    final batchDelete = FirebaseFirestore.instance.batch();
 
-    final querySnapshot = await Firestore.instance
+    final querySnapshot = await FirebaseFirestore.instance
         .collection(FirestorePath.todos(uid))
         .where('complete', isEqualTo: true)
-        .getDocuments();
+        .get();
 
-    for (DocumentSnapshot ds in querySnapshot.documents) {
+    for (DocumentSnapshot ds in querySnapshot.docs) {
       batchDelete.delete(ds.reference);
     }
     await batchDelete.commit();
