@@ -1,33 +1,32 @@
-import 'package:meta/meta.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 
 /*
-This class represent all possible CRUD operation for Firestore.
+This class represent all possible CRUD operation for FirebaseFirestore.
 It contains all generic implementation needed based on the provided document
-path and documentID,since most of the time in Firestore design, we will have
+path and documentID,since most of the time in FirebaseFirestore design, we will have
 documentID and path for any document and collections.
  */
 class FirestoreService {
   FirestoreService._();
   static final instance = FirestoreService._();
 
-  Future<void> setData({
-    @required String path,
-    @required Map<String, dynamic> data,
+  Future<void> set({
+    required String path,
+    required Map<String, dynamic> data,
     bool merge = false,
   }) async {
-    final reference = Firestore.instance.document(path);
+    final reference = FirebaseFirestore.instance.doc(path);
     print('$path: $data');
-    await reference.setData(data, merge: merge);
+    await reference.set(data);
   }
 
   Future<void> bulkSet({
-    @required String path,
-    @required List<Map<String, dynamic>> datas,
+    required String path,
+    required List<Map<String, dynamic>> datas,
     bool merge = false,
   }) async {
-    final reference = Firestore.instance.document(path);
-    final batchSet = Firestore.instance.batch();
+    final reference = FirebaseFirestore.instance.doc(path);
+    final batchSet = FirebaseFirestore.instance.batch();
 
 //    for()
 //    batchSet.
@@ -35,26 +34,27 @@ class FirestoreService {
     print('$path: $datas');
   }
 
-  Future<void> deleteData({@required String path}) async {
-    final reference = Firestore.instance.document(path);
+  Future<void> deleteData({required String path}) async {
+    final reference = FirebaseFirestore.instance.doc(path);
     print('delete: $path');
     await reference.delete();
   }
 
   Stream<List<T>> collectionStream<T>({
-    @required String path,
-    @required T builder(Map<String, dynamic> data, String documentID),
-    Query queryBuilder(Query query),
-    int sort(T lhs, T rhs),
+    required String path,
+    required T builder(Map<String, dynamic> data, String documentID),
+    Query queryBuilder(Query query)?,
+    int sort(T lhs, T rhs)?,
   }) {
-    Query query = Firestore.instance.collection(path);
+    Query query = FirebaseFirestore.instance.collection(path);
     if (queryBuilder != null) {
       query = queryBuilder(query);
     }
     final Stream<QuerySnapshot> snapshots = query.snapshots();
     return snapshots.map((snapshot) {
-      final result = snapshot.documents
-          .map((snapshot) => builder(snapshot.data, snapshot.documentID))
+      final result = snapshot.docs
+          .map((snapshot) =>
+              builder(snapshot.data() as Map<String, dynamic>, snapshot.id))
           .where((value) => value != null)
           .toList();
       if (sort != null) {
@@ -65,12 +65,12 @@ class FirestoreService {
   }
 
   Stream<T> documentStream<T>({
-    @required String path,
-    @required T builder(Map<String, dynamic> data, String documentID),
+    required String path,
+    required T builder(Map<String, dynamic> data, String documentID),
   }) {
-    final DocumentReference reference = Firestore.instance.document(path);
+    final DocumentReference reference = FirebaseFirestore.instance.doc(path);
     final Stream<DocumentSnapshot> snapshots = reference.snapshots();
-    return snapshots
-        .map((snapshot) => builder(snapshot.data, snapshot.documentID));
+    return snapshots.map((snapshot) =>
+        builder(snapshot.data() as Map<String, dynamic>, snapshot.id));
   }
 }
