@@ -1,5 +1,6 @@
 import 'dart:async';
 import 'dart:convert';
+
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
@@ -11,13 +12,14 @@ class AppLocalizations {
   // Helper method to keep the code in the widgets concise
   // Localizations are accessed using an InheritedWidget "of" syntax
   static AppLocalizations of(BuildContext context) {
-    return Localizations.of<AppLocalizations>(context, AppLocalizations);
+    return Localizations.of<AppLocalizations>(context, AppLocalizations)!;
   }
 
   //This is the static member for allowing simple access to the delegate from the MaterialApp
-  static const LocalizationsDelegate<AppLocalizations> delegate = _AppLocalizationsDelegate();
+  static const LocalizationsDelegate<AppLocalizations> delegate =
+      _AppLocalizationsDelegate();
 
-  Map<String, String> _localizedStrings;
+  Map<String, String> _localizedStrings = Map();
 
   Future<bool> load() async {
     // Load the language JSON file from the "lang" folder
@@ -34,9 +36,8 @@ class AppLocalizations {
 
   // This method will be called from every widgets which needs a localized text
   String translate(String key) {
-    return _localizedStrings[key]??key;
+    return _localizedStrings[key] ?? key;
   }
-
 }
 
 // LocalizationsDelegate is a factory for a set of localized resources
