@@ -1,5 +1,5 @@
 /*
-This class defines all the possible read/write locations from the Firestore database.
+This class defines all the possible read/write locations from the FirebaseFirestore database.
 In future, any new path can be added here.
 This class work together with FirestoreService and FirestoreDatabase.
  */
