@@ -25,9 +25,9 @@ class _SplashScreenState extends State<SplashScreen> {
       children: <Widget>[
         Center(
             child: Text(
-              AppLocalizations.of(context).translate("splashTitle"),
+          AppLocalizations.of(context).translate("splashTitle"),
           style: TextStyle(
-            fontSize: Theme.of(context).textTheme.headline.fontSize,
+            fontSize: Theme.of(context).textTheme.headline!.fontSize,
           ),
         )),
         FlutterLogo(
